requirements.txt
python-telegram-bot
Flask
yt-dlp
gunicorn # WSGI server to run Flask reliably